flask/requirements.txt
Flask==1.1.1
Flask-AppBuilder==2.2.0
Flask-Babel==0.12.2
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
importlib-metadata==1.0.0
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.2.0
MarkupSafe==1.1.1
marshmallow==2.19.5
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.19.0
more-itertools==8.0.0
mysqlclient==1.4.6
prison==0.1.2
PyJWT==1.7.1
pyrsistent==0.15.6
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.1.2
six==1.13.0
SQLAlchemy==1.3.11
SQLAlchemy-Utils==0.35.0
Werkzeug==0.16.0
WTForms==2.2.1
zipp==0.6.0
mysql-connector-python